refactor(Lotto): domain 패키지 추가 후 Lotto 클래스 위치 변경
