+++ src/test/java/lotto/domain/LottoTest.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import java.util.stream.Stream;
-import lotto.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
